requirements.txt
Flask==2.3.3
flask-pymongo==2.3.0
Flask-JWT-Extended==4.4.4
Flask-Swagger==0.2.14
python-dotenv==1.0.0
PyMongo==4.5.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
flask-migrate==4.0.4
